Refactor GridLib to use `int64_t` for `Size` and `Index`, replace `grid_to_model/model_to_grid` with `grid_to_world/world_to_grid`, and consolidate `elevations_` to `values_` for consistency. Update tests and logic accordingly.

+++ include/GridLib/Index.hpp
@@ -12,8 +12,8 @@ namespace GridLib
 {
     struct Index
     {
-        size_t row = 0;
-        size_t col = 0;
+        int64_t row = 0;
+        int64_t col = 0;
     };
 
     inline bool operator==(const Index& a, const Index& b)
